refactor(bootstrapper): convert bootstrapper.py to a subpackage
Split the 2,351-line `bootstrapper.py` into a subpackage
`src/fromager/bootstrapper/` with one module per logical layer:

- `_types.py`: type aliases, dataclasses, enums, and constants
- `_cache.py`: shared wheel-cache and prebuilt-download helpers
- `_work_item.py`: `WorkItem` dataclass
- `_phase_item.py`: `PhaseItem` ABC
- `_resolve_item.py`: `ResolveItem` + `_bg_resolve`
- `_start_item.py`: `StartItem`
- `_prepare_source_item.py`: `PrepareSourceItem` + `_bg_prepare_source`
- `_prepare_build_item.py`: `PrepareBuildItem`
- `_build_item.py`: `BuildItem`
- `_process_install_deps_item.py`: `ProcessInstallDepsItem` + `_get_install_dependencies`
- `_complete_item.py`: `CompleteItem`
- `_bootstrapper.py`: `Bootstrapper` class
- `__init__.py`: exports `Bootstrapper` and `_DEFAULT_BG_THREADS` only

Phase modules that call `_cache` helpers use module-level imports
(`from . import _cache`) so mock patch targets remain unambiguous
(e.g. `fromager.bootstrapper._cache._find_cached_wheel`).

Tests updated to import directly from submodules and use the new
patch paths.
