+++ src/fromager/bootstrapper/__init__.py
@@ -1,0 +1,4 @@
+from ._bootstrapper import Bootstrapper
+from ._types import _DEFAULT_BG_THREADS
+
+__all__ = ["_DEFAULT_BG_THREADS", "Bootstrapper"]
