+++ src/fromager/bootstrapper/_bootstrapper.py
@@ -1,1159 +1,51 @@
 from __future__ import annotations
 
-import abc
 import concurrent.futures
 import contextlib
-import dataclasses
 import datetime
 import json
 import logging
 import operator
-import os
 import pathlib
 import shutil
 import tempfile
 import typing
-import zipfile
-from enum import StrEnum
-from urllib.parse import urlparse
 
-import requests.exceptions
 from packaging.requirements import Requirement
 from packaging.utils import NormalizedName, canonicalize_name
 from packaging.version import Version
-from resolvelib.resolvers import ResolverException
 
-from . import (
+from .. import (
     bootstrap_requirement_resolver,
     build_environment,
     dependencies,
-    finders,
     gitutils,
-    hooks,
     progress,
-    resolver,
-    server,
     sources,
-    threading_utils,
-    wheels,
 )
-from .dependency_graph import DependencyGraph
-from .log import req_ctxvar_context, requirement_ctxvar
-from .requirements_file import RequirementType, SourceType
+from ..dependency_graph import DependencyGraph
+from ..log import req_ctxvar_context, requirement_ctxvar
+from ..requirements_file import RequirementType, SourceType
+from . import _cache
+from ._build_item import BuildItem
+from ._phase_item import PhaseItem
+from ._prepare_build_item import PrepareBuildItem
+from ._prepare_source_item import PrepareSourceItem
+from ._process_install_deps_item import ProcessInstallDepsItem
+from ._resolve_item import ResolveItem
+from ._types import (
+    _DEFAULT_BG_THREADS,
+    FailureRecord,
+    FailureType,
+    SeenKey,
+    SourceBuildResult,
+)
+from ._work_item import WorkItem
 
 if typing.TYPE_CHECKING:
-    from . import context
+    from .. import context
 
 logger = logging.getLogger(__name__)
-
-# package name, extras, version, sdist/wheel
-SeenKey = tuple[NormalizedName, tuple[str, ...], str, typing.Literal["sdist", "wheel"]]
-
-_DEFAULT_BG_THREADS: int = max(1, threading_utils.get_cpu_count() // 2)
-
-
-@dataclasses.dataclass
-class SourceBuildResult:
-    """Result of building or downloading a package.
-
-    Captures the output artifacts from either a source build or
-    prebuilt wheel download, used across bootstrap phases.
-    """
-
-    wheel_filename: pathlib.Path | None
-    sdist_filename: pathlib.Path | None
-    unpack_dir: pathlib.Path
-    sdist_root_dir: pathlib.Path | None
-    build_env: build_environment.BuildEnvironment | None
-    source_type: SourceType
-
-
-@dataclasses.dataclass
-class PreparedSourceData:
-    """Result of background I/O pre-fetching returned to the main thread.
-
-    Fields are set in one of three combinations depending on the result type:
-
-    - Source (no cache hit): only ``sdist_root_dir`` is set.
-    - Source (cache hit): both ``sdist_root_dir`` and ``cached_wheel_filename`` are set.
-    - Prebuilt wheel: both ``wheel_filename`` and ``unpack_dir`` are set.
-    """
-
-    # Source path: set after download+unpack OR cache hit
-    sdist_root_dir: pathlib.Path | None = None
-    # Source path: set when the result came from the wheel cache
-    cached_wheel_filename: pathlib.Path | None = None
-    # Prebuilt path: downloaded wheel file
-    wheel_filename: pathlib.Path | None = None
-    # Prebuilt path: unpack directory (created by mkdir)
-    unpack_dir: pathlib.Path | None = None
-
-
-# Valid failure types for test mode error recording
-FailureType = typing.Literal["resolution", "bootstrap", "hook", "dependency_extraction"]
-
-
-class FailureRecord(typing.TypedDict):
-    """Record of a package that failed during bootstrap in test mode.
-
-    Attributes:
-        package: The package name that failed.
-        version: The resolved version (None if resolution failed).
-        exception_type: The exception class name.
-        exception_message: The exception message string.
-        failure_type: Category of failure for analysis.
-    """
-
-    package: str
-    version: str | None
-    exception_type: str
-    exception_message: str
-    failure_type: FailureType
-
-
-class BootstrapPhase(StrEnum):
-    """Processing phases for iterative bootstrap.
-
-    All packages: RESOLVE -> START -> ...
-    Source packages: ... -> PREPARE_SOURCE -> PREPARE_BUILD -> BUILD
-                     -> PROCESS_INSTALL_DEPS -> COMPLETE.
-    Prebuilt packages: ... -> PREPARE_SOURCE -> PROCESS_INSTALL_DEPS -> COMPLETE.
-    """
-
-    RESOLVE = "resolve"
-    START = "start"
-    PREPARE_SOURCE = "prepare-source"
-    PREPARE_BUILD = "prepare-build"
-    BUILD = "build"
-    PROCESS_INSTALL_DEPS = "process-install-deps"
-    COMPLETE = "complete"
-
-
-@dataclasses.dataclass
-class WorkItem:
-    """A unit of work in the iterative bootstrap loop.
-
-    Carries identity fields set at creation time and accumulated state
-    populated across phases as processing advances. The current phase is
-    encoded by the ``PhaseItem`` subclass wrapping this object.
-
-    Items enter at the RESOLVE phase with only req and req_type set.
-    The RESOLVE phase populates source_url and resolved_version, then
-    creates new items at the START phase for each resolved version.
-    """
-
-    # Identity (set at creation)
-    req: Requirement
-    req_type: RequirementType
-    why_snapshot: list[tuple[RequirementType, Requirement, Version]]
-    parent: tuple[Requirement, Version] | None = None
-
-    # Populated by RESOLVE phase (None until then)
-    source_url: str | None = None
-    resolved_version: Version | None = None
-
-    build_sdist_only: bool = False
-
-    # Accumulated state (populated during phases)
-    build_env: build_environment.BuildEnvironment | None = None
-    sdist_root_dir: pathlib.Path | None = None
-    unpack_dir: pathlib.Path | None = None
-    cached_wheel_filename: pathlib.Path | None = None
-    build_result: SourceBuildResult | None = None
-    pbi_pre_built: bool = False
-    exclusive_build: bool = False
-    build_system_deps: set[Requirement] = dataclasses.field(default_factory=set)
-    build_backend_deps: set[Requirement] = dataclasses.field(default_factory=set)
-    build_sdist_deps: set[Requirement] = dataclasses.field(default_factory=set)
-
-    def is_build_requirement_context(self) -> bool:
-        """Return True if this item is being processed as part of a build requirement.
-
-        A package is a build dependency if its own requirement type is
-        build_system, build_backend, or build_sdist, OR if it is an install
-        requirement of something that is itself a build dependency (checked
-        by walking the ``why_snapshot`` chain).
-        """
-        if self.req_type.is_build_requirement:
-            logger.debug(f"is itself a build requirement: {self.req_type}")
-            return True
-        if not self.req_type.is_install_requirement:
-            logger.debug(
-                "is not an install requirement, not checking dependency chain for a build requirement"
-            )
-            return False
-        for req_type, req, resolved_version in reversed(self.why_snapshot):
-            if req_type.is_build_requirement:
-                logger.debug(
-                    f"is a build requirement because {req_type} dependency {req} ({resolved_version}) depends on it"
-                )
-                return True
-        logger.debug("is not a build requirement")
-        return False
-
-
-def _create_unpack_dir(
-    work_dir: pathlib.Path,
-    req: Requirement,
-    resolved_version: Version,
-) -> pathlib.Path:
-    unpack_dir = work_dir / f"{req.name}-{resolved_version}"
-    unpack_dir.mkdir(parents=True, exist_ok=True)
-    return unpack_dir
-
-
-def _extract_build_reqs_from_wheel(
-    work_dir: pathlib.Path,
-    req: Requirement,
-    resolved_version: Version,
-    wheel_filename: pathlib.Path,
-) -> pathlib.Path | None:
-    """Extract fromager build requirement files from a wheel archive.
-
-    Looks for files prefixed with `FROMAGER_BUILD_REQ_PREFIX` inside the
-    wheel's `.dist-info` directory and extracts them to a local unpack
-    directory. Returns the unpack directory on success, or None if the
-    files are not present or extraction fails.
-    """
-    dist_name, dist_version, _, _ = wheels.extract_info_from_wheel_file(
-        req, wheel_filename
-    )
-    unpack_dir = _create_unpack_dir(work_dir, req, resolved_version)
-    dist_filename = f"{dist_name}-{dist_version}"
-    dist_info_path = pathlib.Path(f"{dist_filename}.dist-info")
-    req_filenames: list[str] = [
-        dependencies.BUILD_BACKEND_REQ_FILE_NAME,
-        dependencies.BUILD_SDIST_REQ_FILE_NAME,
-        dependencies.BUILD_SYSTEM_REQ_FILE_NAME,
-    ]
-    try:
-        with zipfile.ZipFile(wheel_filename) as archive:
-            for filename in req_filenames:
-                zipinfo = archive.getinfo(
-                    str(
-                        dist_info_path
-                        / f"{wheels.FROMAGER_BUILD_REQ_PREFIX}-{filename}"
-                    )
-                )
-                if os.path.isabs(zipinfo.filename) or ".." in zipinfo.filename:
-                    raise ValueError(f"Unsafe path in wheel: {zipinfo.filename}")
-                zipinfo.filename = filename
-                output_file = archive.extract(zipinfo, unpack_dir)
-                logger.info(f"extracted {output_file}")
-        logger.info(f"extracted build requirements from wheel into {unpack_dir}")
-        return unpack_dir
-    except Exception as e:
-        logger.info(f"could not extract build requirements from wheel: {e}")
-        for filename in req_filenames:
-            unpack_dir.joinpath(filename).unlink(missing_ok=True)
-        return None
-
-
-def _look_for_existing_wheel(
-    ctx: context.WorkContext,
-    req: Requirement,
-    resolved_version: Version,
-    search_in: pathlib.Path,
-) -> tuple[pathlib.Path | None, pathlib.Path | None]:
-    pbi = ctx.package_build_info(req)
-    expected_build_tag = pbi.build_tag(resolved_version)
-    logger.info(
-        f"looking for existing wheel for version {resolved_version} with build tag {expected_build_tag} in {search_in}"
-    )
-    wheel_filename = finders.find_wheel(
-        downloads_dir=search_in,
-        req=req,
-        dist_version=str(resolved_version),
-        build_tag=expected_build_tag,
-    )
-    if not wheel_filename:
-        return None, None
-    _, _, build_tag, _ = wheels.extract_info_from_wheel_file(req, wheel_filename)
-    if expected_build_tag and expected_build_tag != build_tag:
-        logger.info(
-            f"found wheel for {resolved_version} in {wheel_filename} but build tag does not match. Got {build_tag} but expected {expected_build_tag}"
-        )
-        return None, None
-    logger.info(f"found existing wheel {wheel_filename}")
-    build_reqs_dir = _extract_build_reqs_from_wheel(
-        ctx.work_dir, req, resolved_version, wheel_filename
-    )
-    return wheel_filename, build_reqs_dir
-
-
-def _download_wheel_from_cache(
-    ctx: context.WorkContext,
-    cache_wheel_server_url: str | None,
-    req: Requirement,
-    resolved_version: Version,
-) -> tuple[pathlib.Path | None, pathlib.Path | None]:
-    if not cache_wheel_server_url:
-        return None, None
-    logger.info(f"checking if wheel was already uploaded to {cache_wheel_server_url}")
-    try:
-        pinned_req = Requirement(f"{req.name}=={resolved_version}")
-        provider = finders.PyPICacheProvider(
-            cache_server_url=cache_wheel_server_url,
-            constraints=ctx.constraints,
-        )
-        results = resolver.find_all_matching_from_provider(provider, pinned_req)
-        wheel_url, _ = results[0]
-        wheelfile_name = pathlib.Path(urlparse(wheel_url).path)
-        pbi = ctx.package_build_info(req)
-        expected_build_tag = pbi.build_tag(resolved_version)
-        logger.info(f"has expected build tag {expected_build_tag}")
-        changelogs = pbi.get_changelog(resolved_version)
-        logger.debug(f"has change logs {changelogs}")
-
-        _, _, build_tag, _ = wheels.extract_info_from_wheel_file(req, wheelfile_name)
-        if expected_build_tag and expected_build_tag != build_tag:
-            logger.info(
-                f"found wheel for {resolved_version} in cache but build tag does not match. Got {build_tag} but expected {expected_build_tag}"
-            )
-            return None, None
-
-        cached_wheel = wheels.download_wheel(
-            req=req, wheel_url=wheel_url, output_directory=ctx.wheels_downloads
-        )
-        if cache_wheel_server_url != ctx.wheel_server_url:
-            server.update_wheel_mirror(ctx)
-        logger.info("found built wheel on cache server")
-        unpack_dir = _extract_build_reqs_from_wheel(
-            ctx.work_dir, req, resolved_version, cached_wheel
-        )
-        return cached_wheel, unpack_dir
-    except ResolverException:
-        logger.info(
-            f"did not find wheel for {resolved_version} in {cache_wheel_server_url}"
-        )
-        return None, None
-    except requests.exceptions.RequestException as err:
-        logger.warning(
-            f"network error checking wheel cache for {resolved_version} "
-            f"at {cache_wheel_server_url}: {err}"
-        )
-        return None, None
-    except Exception as err:
-        logger.warning(
-            f"unexpected error checking wheel cache for {resolved_version} "
-            f"at {cache_wheel_server_url}: {err}"
-        )
-        return None, None
-
-
-def _find_cached_wheel(
-    ctx: context.WorkContext,
-    cache_wheel_server_url: str | None,
-    req: Requirement,
-    resolved_version: Version,
-) -> tuple[pathlib.Path | None, pathlib.Path | None]:
-    """Look for cached wheel in 3 locations (thread-safe, no Bootstrapper state).
-
-    Checks for cached wheels in order:
-    1. wheels_build directory (previously built)
-    2. wheels_downloads directory (previously downloaded)
-    3. Cache server (remote cache)
-
-    Returns:
-        Tuple of (cached_wheel_filename, unpacked_cached_wheel).
-        Both None if no cache hit.
-    """
-    cached_wheel, unpacked = _look_for_existing_wheel(
-        ctx, req, resolved_version, ctx.wheels_build
-    )
-    if cached_wheel:
-        return cached_wheel, unpacked
-
-    cached_wheel, unpacked = _look_for_existing_wheel(
-        ctx, req, resolved_version, ctx.wheels_downloads
-    )
-    if cached_wheel:
-        return cached_wheel, unpacked
-
-    cached_wheel, unpacked = _download_wheel_from_cache(
-        ctx, cache_wheel_server_url, req, resolved_version
-    )
-    if cached_wheel:
-        return cached_wheel, unpacked
-
-    return None, None
-
-
-def _bg_resolve(
-    bg_resolver: bootstrap_requirement_resolver.BootstrapRequirementResolver,
-    req: Requirement,
-    req_type: RequirementType,
-    parent_req: Requirement | None,
-    return_all_versions: bool,
-) -> list[tuple[str, Version]]:
-    """Background-safe resolution: no Bootstrapper state accessed."""
-    logger.info(f"{BootstrapPhase.RESOLVE} for {req_type} requirement")
-    return bg_resolver.resolve(
-        req=req,
-        req_type=req_type,
-        parent_req=parent_req,
-        return_all_versions=return_all_versions,
-    )
-
-
-def _bg_prepare_source(
-    ctx: context.WorkContext,
-    cache_wheel_server_url: str | None,
-    req: Requirement,
-    resolved_version: Version,
-    source_url: str,
-) -> PreparedSourceData:
-    """Background-safe source download+unpack: no Bootstrapper state accessed."""
-    # Thread-safe: _seen_requirements in the main thread prevents the same
-    # package from being submitted to the thread pool more than once.
-    # Paths from download_source() and prepare_source() include {name}-{version},
-    # making them unique across concurrent threads processing different packages.
-    logger.info("preparing source")
-    cached_wheel, unpacked = _find_cached_wheel(
-        ctx, cache_wheel_server_url, req, resolved_version
-    )
-    if unpacked is not None:
-        return PreparedSourceData(
-            sdist_root_dir=unpacked / unpacked.stem,
-            cached_wheel_filename=cached_wheel,
-        )
-    source_filename = sources.download_source(
-        ctx=ctx, req=req, version=resolved_version, download_url=source_url
-    )
-    sdist_root_dir = sources.prepare_source(
-        ctx=ctx, req=req, source_filename=source_filename, version=resolved_version
-    )
-    return PreparedSourceData(sdist_root_dir=sdist_root_dir)
-
-
-def _bg_prepare_prebuilt(
-    ctx: context.WorkContext,
-    req: Requirement,
-    req_type: RequirementType,
-    resolved_version: Version,
-    wheel_url: str,
-) -> PreparedSourceData:
-    """Background-safe prebuilt download: no Bootstrapper state accessed."""
-    # Thread-safe: paths include {req.name}-{resolved_version} (unique per package),
-    # mkdir uses exist_ok=True (atomic), and update_wheel_mirror() is already locked.
-    logger.info(f"using pre-built wheel for {req_type} requirement")
-    wheel_filename = wheels.download_wheel(req, wheel_url, ctx.wheels_prebuilt)
-    unpack_dir = ctx.work_dir / f"{req.name}-{resolved_version}"
-    unpack_dir.mkdir(parents=True, exist_ok=True)
-    server.update_wheel_mirror(ctx)
-    return PreparedSourceData(wheel_filename=wheel_filename, unpack_dir=unpack_dir)
-
-
-def _get_install_dependencies(
-    ctx: context.WorkContext,
-    req: Requirement,
-    resolved_version: Version,
-    wheel_filename: pathlib.Path | None,
-    sdist_filename: pathlib.Path | None,
-    sdist_root_dir: pathlib.Path | None,
-    build_env: build_environment.BuildEnvironment | None,
-    unpack_dir: pathlib.Path | None,
-) -> list[Requirement]:
-    """Extract install dependencies from a built wheel or sdist.
-
-    Returns:
-        List of install requirements.
-
-    Raises:
-        RuntimeError: If both wheel_filename and sdist_filename are None.
-    """
-    if wheel_filename is not None:
-        assert unpack_dir is not None
-        logger.debug(
-            "get install dependencies of wheel %s",
-            wheel_filename.name,
-        )
-        return list(
-            dependencies.get_install_dependencies_of_wheel(
-                req=req,
-                wheel_filename=wheel_filename,
-                requirements_file_dir=unpack_dir,
-            )
-        )
-    elif sdist_filename is not None:
-        assert sdist_root_dir is not None
-        assert build_env is not None
-        logger.debug(
-            "get install dependencies of sdist from directory %s",
-            sdist_root_dir,
-        )
-        return list(
-            dependencies.get_install_dependencies_of_sdist(
-                ctx=ctx,
-                req=req,
-                version=resolved_version,
-                sdist_root_dir=sdist_root_dir,
-                build_env=build_env,
-            )
-        )
-    else:
-        raise RuntimeError("wheel_filename and sdist_filename are None")
-
-
-class PhaseItem(abc.ABC):
-    """Abstract base for items pushed onto the bootstrap stack.
-
-    Each subclass encodes one phase of the bootstrap pipeline.
-    Wraps a ``WorkItem`` (accumulated per-package state) and implements
-    the processing logic for that phase in ``run()``.
-    """
-
-    phase: typing.ClassVar[BootstrapPhase]
-    tracks_why: typing.ClassVar[bool] = True
-
-    def __init__(self, work_item: WorkItem) -> None:
-        self.work_item = work_item
-        self.bg_future: concurrent.futures.Future[typing.Any] | None = None
-
-    @abc.abstractmethod
-    def run(self, bt: Bootstrapper) -> list[PhaseItem]: ...
-
-    @property
-    def requires_exclusive_run(self) -> bool:
-        """Return True if this item must run without concurrent background I/O.
-
-        When True, the bootstrap loop drains the background thread pool before
-        calling ``run()``. Override in subclasses that require exclusive access.
-        """
-        return False
-
-    def background_work(
-        self, bt: Bootstrapper
-    ) -> typing.Callable[[], typing.Any] | None:
-        """Return a zero-argument callable for background I/O, or None.
-
-        Override in subclasses that need background prefetching.
-        ``bt`` is provided so subclasses can capture Bootstrapper state
-        (e.g. resolver, ctx) into the returned closure without storing
-        a circular reference on the item itself.
-        """
-        return None
-
-    def __str__(self) -> str:
-        """Human-readable representation: ``"<ClassName>(<req>)"``."""
-        wi = self.work_item
-        return f"{type(self).__name__}({wi.req})"
-
-    def as_json(self) -> dict[str, typing.Any]:
-        """Return a JSON-serialisable dict for stack-state recording."""
-        wi = self.work_item
-        return {
-            "req": str(wi.req),
-            "req_type": str(wi.req_type),
-            "phase": str(self.phase),
-            "resolved_version": str(wi.resolved_version)
-            if wi.resolved_version is not None
-            else None,
-            "source_url": wi.source_url,
-            "build_sdist_only": wi.build_sdist_only,
-            "why": [
-                {"req_type": str(rt), "req": str(r), "version": str(v)}
-                for rt, r, v in wi.why_snapshot
-            ],
-            "parent": (
-                {"req": str(wi.parent[0]), "version": str(wi.parent[1])}
-                if wi.parent
-                else None
-            ),
-            "build_system_deps": sorted(str(r) for r in wi.build_system_deps),
-            "build_backend_deps": sorted(str(r) for r in wi.build_backend_deps),
-            "build_sdist_deps": sorted(str(r) for r in wi.build_sdist_deps),
-        }
-
-
-class ResolveItem(PhaseItem):
-    """RESOLVE phase: resolve versions and expand into StartItems."""
-
-    phase: typing.ClassVar[BootstrapPhase] = BootstrapPhase.RESOLVE
-    tracks_why: typing.ClassVar[bool] = False
-
-    def background_work(
-        self, bt: Bootstrapper
-    ) -> typing.Callable[[], typing.Any] | None:
-        """Return closure that calls ``_bg_resolve`` in a thread."""
-        bg_resolver = bt.resolver
-        req = self.work_item.req
-        req_type = self.work_item.req_type
-        parent_req = (
-            self.work_item.why_snapshot[-1][1] if self.work_item.why_snapshot else None
-        )
-        return_all = bt.multiple_versions
-
-        def do_resolve() -> list[tuple[str, Version]]:
-            with req_ctxvar_context(req):
-                return _bg_resolve(bg_resolver, req, req_type, parent_req, return_all)
-
-        return do_resolve
-
-    def run(self, bt: Bootstrapper) -> list[PhaseItem]:
-        """RESOLVE phase: resolve versions and expand into StartItems.
-
-        Centralizes version resolution so all dependencies are expanded
-        uniformly. In multiple_versions mode, filters out versions that
-        already failed in this run and versions whose wheels are already
-        cached to avoid redundant builds and transitive dependency
-        processing.
-
-        Returns:
-            One StartItem per resolved version that needs building.
-            Empty list if all versions are already cached.
-        """
-        assert self.bg_future is not None
-        # bg_future.result() blocks until the background resolution completes,
-        # then returns the result or re-raises any exception from the background.
-        resolved_versions = self.bg_future.result()
-        if not resolved_versions:
-            raise RuntimeError(
-                f"Could not resolve any versions for {self.work_item.req}"
-            )
-
-        if bt.multiple_versions:
-            pkg_name = canonicalize_name(self.work_item.req.name)
-            resolved_versions = [
-                (url, ver)
-                for url, ver in resolved_versions
-                if not bt.has_failed_version(pkg_name, ver)
-            ]
-            if not resolved_versions:
-                raise RuntimeError(
-                    f"Could not resolve any versions for {self.work_item.req}"
-                    f" (all candidates failed previously)"
-                )
-
-            logger.info(
-                f"resolved {len(resolved_versions)} version(s) for {self.work_item.req}"
-            )
-            filtered: list[tuple[str, Version]] = []
-            for source_url, version in resolved_versions:
-                cached_wheel, _ = _find_cached_wheel(
-                    bt.ctx, bt.cache_wheel_server_url, self.work_item.req, version
-                )
-                if cached_wheel:
-                    logger.info(
-                        f"{self.work_item.req.name}=={version}: wheel already cached "
-                        f"at {cached_wheel.name}, skipping"
-                    )
-                else:
-                    filtered.append((source_url, version))
-            if not filtered:
-                # Always process the highest version (first in
-                # resolved_versions) so new transitive dependencies
-                # are discovered even when every wheel is cached.
-                logger.info(
-                    f"all versions of {self.work_item.req.name} already cached, "
-                    f"keeping highest version {resolved_versions[0][1]} "
-                    f"for dependency discovery"
-                )
-                filtered.append(resolved_versions[0])
-            resolved_versions = filtered
-
-        # Build list so highest version ends up on top of the stack
-        # (last element after extend) and is processed first.
-        items: list[PhaseItem] = []
-        for source_url, version in reversed(resolved_versions):
-            items.append(
-                StartItem(
-                    WorkItem(
-                        req=self.work_item.req,
-                        req_type=self.work_item.req_type,
-                        why_snapshot=list(self.work_item.why_snapshot),
-                        parent=self.work_item.parent,
-                        source_url=source_url,
-                        resolved_version=version,
-                    )
-                )
-            )
-        return items
-
-
-class StartItem(PhaseItem):
-    """START phase: add to graph, check if already seen."""
-
-    phase: typing.ClassVar[BootstrapPhase] = BootstrapPhase.START
-    tracks_why: typing.ClassVar[bool] = False
-
-    def run(self, bt: Bootstrapper) -> list[PhaseItem]:
-        """START phase: add to graph, check if already seen.
-
-        _track_why is a no-op for this phase (tracks_why is False),
-        matching the original behavior where graph addition and
-        seen-check happen before pushing onto the why stack.
-
-        Returns:
-            Empty list if already seen (nothing to do).
-            [PrepareSourceItem] if this is new work.
-        """
-        wi = self.work_item
-        assert wi.resolved_version is not None
-        assert wi.source_url is not None
-
-        # Add to graph (skip TOP_LEVEL, already added in _resolve_and_add_top_level)
-        if wi.req_type != RequirementType.TOP_LEVEL:
-            bt.add_to_graph(
-                wi.req,
-                wi.req_type,
-                wi.resolved_version,
-                wi.source_url,
-                wi.parent,
-            )
-
-        wi.build_sdist_only = bt.sdist_only and not wi.is_build_requirement_context()
-
-        if bt.has_been_seen(wi.req, wi.resolved_version, wi.build_sdist_only):
-            logger.debug(
-                f"redundant {wi.req_type} dependency {wi.req} "
-                f"({wi.resolved_version}, sdist_only={wi.build_sdist_only}) "
-                f"for {bt.explain}"
-            )
-            return []
-        bt.mark_as_seen(wi.req, wi.resolved_version, wi.build_sdist_only)
-
-        logger.info(
-            f"new {wi.req_type} dependency {wi.req} resolves to {wi.resolved_version}"
-        )
-
-        # Must set pbi_pre_built before constructing PrepareSourceItem so that
-        # PrepareSourceItem.background_work() immediately sees the correct value.
-        pbi = bt.ctx.package_build_info(wi.req)
-        wi.pbi_pre_built = pbi.pre_built
-        wi.exclusive_build = pbi.exclusive_build
-        return [PrepareSourceItem(wi)]
-
-
-class PrepareSourceItem(PhaseItem):
-    """PREPARE_SOURCE phase: download source or prebuilt, get build system deps."""
-
-    phase: typing.ClassVar[BootstrapPhase] = BootstrapPhase.PREPARE_SOURCE
-    tracks_why: typing.ClassVar[bool] = True
-
-    def background_work(
-        self, bt: Bootstrapper
-    ) -> typing.Callable[[], typing.Any] | None:
-        """Return closure for background source download or prebuilt fetch."""
-        wi = self.work_item
-        assert wi.resolved_version is not None
-        assert wi.source_url is not None
-        ctx = bt.ctx
-        cache_wheel_server_url = bt.cache_wheel_server_url
-        req = wi.req
-        req_type = wi.req_type
-        resolved_version = wi.resolved_version
-        source_url = wi.source_url
-
-        if wi.pbi_pre_built:
-
-            def do_prepare_prebuilt() -> PreparedSourceData:
-                with req_ctxvar_context(req, resolved_version):
-                    return _bg_prepare_prebuilt(
-                        ctx, req, req_type, resolved_version, source_url
-                    )
-
-            return do_prepare_prebuilt
-
-        def do_prepare_source() -> PreparedSourceData:
-            with req_ctxvar_context(req, resolved_version):
-                return _bg_prepare_source(
-                    ctx, cache_wheel_server_url, req, resolved_version, source_url
-                )
-
-        return do_prepare_source
-
-    def run(self, bt: Bootstrapper) -> list[PhaseItem]:
-        """PREPARE_SOURCE phase: download source or prebuilt, get build system deps.
-
-        Uses background I/O result from ``self.bg_future`` when available,
-        falling back to inline I/O otherwise.
-
-        Returns:
-            Prebuilt: [ProcessInstallDepsItem] (skip build phases).
-            Source: [PrepareBuildItem, *build_system_dep_items].
-        """
-        wi = self.work_item
-        assert wi.resolved_version is not None
-        assert wi.source_url is not None
-
-        # bg_future is always set for PREPARE_SOURCE items (see _push_items).
-        # bg_future.result() blocks until done and re-raises any background exception.
-        assert self.bg_future is not None
-        prepared: PreparedSourceData = self.bg_future.result()
-
-        constraint = bt.ctx.constraints.get_constraint(wi.req.name)
-        if constraint:
-            logger.info(
-                f"incoming requirement {wi.req} matches constraint "
-                f"{constraint}. Will apply both."
-            )
-
-        if wi.pbi_pre_built:
-            # Background task already downloaded the prebuilt wheel
-            assert prepared.wheel_filename is not None
-            assert prepared.unpack_dir is not None
-            wi.build_result = SourceBuildResult(
-                wheel_filename=prepared.wheel_filename,
-                sdist_filename=None,
-                unpack_dir=prepared.unpack_dir,
-                sdist_root_dir=None,
-                build_env=None,
-                source_type=SourceType.PREBUILT,
-            )
-            return [ProcessInstallDepsItem(wi)]
-
-        # Source build path: background task already downloaded and prepared the source
-        assert prepared.sdist_root_dir is not None
-        sdist_root_dir = prepared.sdist_root_dir
-        wi.cached_wheel_filename = prepared.cached_wheel_filename
-
-        assert sdist_root_dir is not None
-
-        if sdist_root_dir.parent.parent != bt.ctx.work_dir:
-            raise ValueError(f"'{sdist_root_dir}/../..' should be {bt.ctx.work_dir}")
-        wi.sdist_root_dir = sdist_root_dir
-        wi.unpack_dir = sdist_root_dir.parent
-
-        wi.build_env = build_environment.BuildEnvironment(
-            ctx=bt.ctx,
-            parent_dir=sdist_root_dir.parent,
-        )
-
-        # Get build system dependencies
-        wi.build_system_deps = dependencies.get_build_system_dependencies(
-            ctx=bt.ctx,
-            req=wi.req,
-            version=wi.resolved_version,
-            sdist_root_dir=sdist_root_dir,
-        )
-
-        dep_items = bt.create_unresolved_work_items(
-            wi.build_system_deps,
-            RequirementType.BUILD_SYSTEM,
-            wi.req,
-            wi.resolved_version,
-        )
-
-        return [PrepareBuildItem(wi)] + dep_items
-
-
-class PrepareBuildItem(PhaseItem):
-    """PREPARE_BUILD phase: install system deps, get backend/sdist deps."""
-
-    phase: typing.ClassVar[BootstrapPhase] = BootstrapPhase.PREPARE_BUILD
-    tracks_why: typing.ClassVar[bool] = True
-
-    def run(self, bt: Bootstrapper) -> list[PhaseItem]:
-        """PREPARE_BUILD phase: install system deps, get backend/sdist deps.
-
-        Build-backend and build-sdist dependencies that are already satisfied
-        by a resolved build-system dependency reuse that version instead of
-        resolving independently (see :issue:`1194`).
-
-        Returns:
-            [BuildItem, *backend_dep_items, *sdist_dep_items].
-        """
-        wi = self.work_item
-        assert wi.resolved_version is not None
-        assert wi.build_env is not None
-        assert wi.sdist_root_dir is not None
-
-        # Install build system deps (their wheels exist from DFS processing)
-        wi.build_env.install(wi.build_system_deps)
-
-        # Get build backend dependencies
-        wi.build_backend_deps = dependencies.get_build_backend_dependencies(
-            ctx=bt.ctx,
-            req=wi.req,
-            version=wi.resolved_version,
-            sdist_root_dir=wi.sdist_root_dir,
-            build_env=wi.build_env,
-        )
-
-        # Get build sdist dependencies
-        wi.build_sdist_deps = dependencies.get_build_sdist_dependencies(
-            ctx=bt.ctx,
-            req=wi.req,
-            version=wi.resolved_version,
-            sdist_root_dir=wi.sdist_root_dir,
-            build_env=wi.build_env,
-        )
-
-        # Filter out deps already satisfied by build-system dependencies
-        # to avoid resolving to a different (typically newer) version.
-        resolved_build_sys = bt.get_resolved_build_system_versions(wi)
-        parent = (wi.req, wi.resolved_version)
-        wi.build_backend_deps = bt.filter_deps_satisfied_by_build_system(
-            wi.build_backend_deps,
-            resolved_build_sys,
-            RequirementType.BUILD_BACKEND,
-            parent,
-        )
-        wi.build_sdist_deps = bt.filter_deps_satisfied_by_build_system(
-            wi.build_sdist_deps,
-            resolved_build_sys,
-            RequirementType.BUILD_SDIST,
-            parent,
-        )
-
-        backend_items = bt.create_unresolved_work_items(
-            wi.build_backend_deps,
-            RequirementType.BUILD_BACKEND,
-            wi.req,
-            wi.resolved_version,
-        )
-        sdist_items = bt.create_unresolved_work_items(
-            wi.build_sdist_deps,
-            RequirementType.BUILD_SDIST,
-            wi.req,
-            wi.resolved_version,
-        )
-        dep_items = backend_items + sdist_items
-
-        return [BuildItem(wi)] + dep_items
-
-
-class BuildItem(PhaseItem):
-    """BUILD phase: install remaining deps, build wheel/sdist."""
-
-    phase: typing.ClassVar[BootstrapPhase] = BootstrapPhase.BUILD
-    tracks_why: typing.ClassVar[bool] = True
-
-    @property
-    def requires_exclusive_run(self) -> bool:
-        return self.work_item.exclusive_build
-
-    def run(self, bt: Bootstrapper) -> list[PhaseItem]:
-        """BUILD phase: install remaining deps, build wheel/sdist.
-
-        Returns:
-            [ProcessInstallDepsItem].
-        """
-        wi = self.work_item
-        assert wi.resolved_version is not None
-        assert wi.build_env is not None
-        assert wi.sdist_root_dir is not None
-
-        # Install backend+sdist deps if disjoint from system deps
-        remaining_deps = wi.build_backend_deps | wi.build_sdist_deps
-        if remaining_deps.isdisjoint(wi.build_system_deps):
-            wi.build_env.install(remaining_deps)
-
-        wheel_filename, sdist_filename = self.do_build(bt.ctx, bt.explain)
-
-        source_type = sources.get_source_type(bt.ctx, wi.req)
-
-        wi.build_result = SourceBuildResult(
-            wheel_filename=wheel_filename,
-            sdist_filename=sdist_filename,
-            unpack_dir=wi.sdist_root_dir.parent,
-            sdist_root_dir=wi.sdist_root_dir,
-            build_env=wi.build_env,
-            source_type=source_type,
-        )
-
-        return [ProcessInstallDepsItem(wi)]
-
-    def _build_sdist(self, ctx: context.WorkContext) -> pathlib.Path:
-        """Build or locate an sdist for this item's package.
-
-        Checks ``ctx.sdists_builds`` for an existing sdist before invoking
-        ``sources.build_sdist()``.
-        """
-        wi = self.work_item
-        assert wi.resolved_version is not None
-        assert wi.sdist_root_dir is not None
-        assert wi.build_env is not None
-        find_sdist_result = finders.find_sdist(
-            ctx, ctx.sdists_builds, wi.req, str(wi.resolved_version)
-        )
-        if not find_sdist_result:
-            sdist_filename: pathlib.Path = sources.build_sdist(
-                ctx=ctx,
-                req=wi.req,
-                version=wi.resolved_version,
-                sdist_root_dir=wi.sdist_root_dir,
-                build_env=wi.build_env,
-            )
-        else:
-            sdist_filename = find_sdist_result
-            logger.info(
-                f"have sdist version {wi.resolved_version}: {find_sdist_result}"
-            )
-        return sdist_filename
-
-    def _build_wheel(
-        self,
-        ctx: context.WorkContext,
-        explain: str = "",
-    ) -> tuple[pathlib.Path, pathlib.Path]:
-        """Build a wheel for this item's package.
-
-        Calls :meth:`_build_sdist` first, then invokes ``wheels.build_wheel()``
-        and updates the local wheel mirror. Returns ``(wheel_filename,
-        sdist_filename)`` where *wheel_filename* is the path in
-        ``ctx.wheels_downloads`` after the mirror update.
-        """
-        wi = self.work_item
-        assert wi.resolved_version is not None
-        assert wi.sdist_root_dir is not None
-        assert wi.build_env is not None
-        sdist_filename = self._build_sdist(ctx)
-
-        logger.info(f"starting build of {explain} for {ctx.variant}")
-        built_filename = wheels.build_wheel(
-            ctx=ctx,
-            req=wi.req,
-            sdist_root_dir=wi.sdist_root_dir,
-            version=wi.resolved_version,
-            build_env=wi.build_env,
-        )
-        server.update_wheel_mirror(ctx)
-        # When we update the mirror, the built file moves to the downloads directory.
-        wheel_filename = ctx.wheels_downloads / built_filename.name
-        logger.info(f"built wheel for version {wi.resolved_version}: {wheel_filename}")
-        return wheel_filename, sdist_filename
-
-    def do_build(
-        self,
-        ctx: context.WorkContext,
-        explain: str = "",
-    ) -> tuple[pathlib.Path | None, pathlib.Path | None]:
-        """Build a wheel or sdist for this item's package.
-
-        Returns ``(wheel_filename, sdist_filename)``.  Either value may be
-        ``None`` depending on the build path:
-
-        - Cached wheel:  ``(cached_wheel_filename, None)``
-        - sdist-only:    ``(None, sdist_filename)``
-        - Full wheel:    ``(wheel_filename, sdist_filename)``
-        """
-        wi = self.work_item
-        if wi.cached_wheel_filename:
-            logger.debug(
-                f"getting install requirements from cached wheel {wi.cached_wheel_filename.name}"
-            )
-            return wi.cached_wheel_filename, None
-        elif wi.build_sdist_only:
-            logger.debug(
-                f"getting install requirements from sdist {wi.req.name}=={wi.resolved_version}"
-            )
-            return None, self._build_sdist(ctx)
-        else:
-            logger.debug(
-                f"building wheel {wi.req.name}=={wi.resolved_version} to get install requirements"
-            )
-            return self._build_wheel(ctx, explain)
-
-
-class ProcessInstallDepsItem(PhaseItem):
-    """PROCESS_INSTALL_DEPS phase: hooks, extract deps, build order."""
-
-    phase: typing.ClassVar[BootstrapPhase] = BootstrapPhase.PROCESS_INSTALL_DEPS
-    tracks_why: typing.ClassVar[bool] = True
-
-    def run(self, bt: Bootstrapper) -> list[PhaseItem]:
-        """PROCESS_INSTALL_DEPS phase: hooks, extract deps, build order.
-
-        Returns:
-            [CompleteItem, *install_dep_items].
-        """
-        wi = self.work_item
-        assert wi.resolved_version is not None
-        assert wi.source_url is not None
-        assert wi.build_result is not None
-
-        # Run post-bootstrap hooks (non-fatal in test mode)
-        try:
-            hooks.run_post_bootstrap_hooks(
-                ctx=bt.ctx,
-                req=wi.req,
-                dist_name=canonicalize_name(wi.req.name),
-                dist_version=str(wi.resolved_version),
-                sdist_filename=wi.build_result.sdist_filename,
-                wheel_filename=wi.build_result.wheel_filename,
-            )
-        except Exception as hook_error:
-            if not bt.test_mode:
-                raise
-            bt.record_test_mode_failure(
-                wi.req,
-                str(wi.resolved_version),
-                hook_error,
-                "hook",
-                "warning",
-            )
-
-        # Extract install dependencies (non-fatal in test mode)
-        try:
-            install_dependencies = _get_install_dependencies(
-                ctx=bt.ctx,
-                req=wi.req,
-                resolved_version=wi.resolved_version,
-                wheel_filename=wi.build_result.wheel_filename,
-                sdist_filename=wi.build_result.sdist_filename,
-                sdist_root_dir=wi.build_result.sdist_root_dir,
-                build_env=wi.build_result.build_env,
-                unpack_dir=wi.build_result.unpack_dir,
-            )
-        except Exception as dep_error:
-            if not bt.test_mode:
-                raise
-            bt.record_test_mode_failure(
-                wi.req,
-                str(wi.resolved_version),
-                dep_error,
-                "dependency_extraction",
-                "warning",
-            )
-            install_dependencies = []
-
-        logger.debug(
-            "install dependencies: %s",
-            ", ".join(sorted(str(r) for r in install_dependencies)),
-        )
-
-        pbi = bt.ctx.package_build_info(wi.req)
-        constraint = bt.ctx.constraints.get_constraint(wi.req.name)
-        bt.add_to_build_order(
-            req=wi.req,
-            version=wi.resolved_version,
-            source_url=wi.source_url,
-            source_type=wi.build_result.source_type,
-            prebuilt=pbi.pre_built,
-            constraint=constraint,
-        )
-
-        dep_items = bt.create_unresolved_work_items(
-            install_dependencies,
-            RequirementType.INSTALL,
-            wi.req,
-            wi.resolved_version,
-        )
-
-        return [CompleteItem(wi)] + dep_items
-
-
-class CompleteItem(PhaseItem):
-    """COMPLETE phase: clean up build directories."""
-
-    phase: typing.ClassVar[BootstrapPhase] = BootstrapPhase.COMPLETE
-    tracks_why: typing.ClassVar[bool] = True
-
-    def run(self, bt: Bootstrapper) -> list[PhaseItem]:
-        """COMPLETE phase: clean up build directories.
-
-        Returns:
-            Empty list (processing finished for this item).
-        """
-        wi = self.work_item
-        if wi.build_result is not None:
-            bt.ctx.clean_build_dirs(
-                wi.build_result.sdist_root_dir,
-                wi.build_result.build_env,
-            )
-        return []
 
 
 class Bootstrapper:
@@ -1664,7 +556,7 @@ class Bootstrapper:
         resolved_version: Version,
         wheel_url: str,
     ) -> tuple[pathlib.Path, pathlib.Path]:
-        result = _bg_prepare_prebuilt(
+        result = _cache._bg_prepare_prebuilt(
             self.ctx, req, req_type, resolved_version, wheel_url
         )
         assert result.wheel_filename is not None
@@ -1756,6 +648,7 @@ class Bootstrapper:
                 req.name,
                 fallback_version,
             )
+
             # Package succeeded via fallback - no failure to record
 
             return SourceBuildResult(
