+++ tests/test_bootstrapper.py
@@ -12,6 +12,28 @@ from packaging.version import Version
 from resolvelib.resolvers import ResolverException
 
 from fromager import bootstrapper, log
+from fromager.bootstrapper._build_item import BuildItem
+from fromager.bootstrapper._cache import (
+    _bg_prepare_prebuilt,
+    _download_wheel_from_cache,
+    _find_cached_wheel,
+)
+from fromager.bootstrapper._phase_item import PhaseItem
+from fromager.bootstrapper._prepare_source_item import (
+    PrepareSourceItem,
+    _bg_prepare_source,
+)
+from fromager.bootstrapper._process_install_deps_item import (
+    ProcessInstallDepsItem,
+    _get_install_dependencies,
+)
+from fromager.bootstrapper._resolve_item import ResolveItem
+from fromager.bootstrapper._start_item import StartItem
+from fromager.bootstrapper._types import (
+    BootstrapPhase,
+    SourceBuildResult,
+)
+from fromager.bootstrapper._work_item import WorkItem
 from fromager.context import WorkContext
 from fromager.requirements_file import RequirementType, SourceType
 
@@ -198,8 +220,8 @@ def test_explain(tmp_context: WorkContext) -> None:
 def _make_work_item(
     req_type: RequirementType,
     why_snapshot: list[tuple[RequirementType, Requirement, Version]] | None = None,
-) -> bootstrapper.WorkItem:
-    return bootstrapper.WorkItem(
+) -> WorkItem:
+    return WorkItem(
         req=Requirement("testpkg"),
         req_type=req_type,
         why_snapshot=why_snapshot or [],
@@ -253,7 +275,7 @@ def test_find_cached_wheel_returns_tuple(tmp_context: WorkContext) -> None:
     bt = bootstrapper.Bootstrapper(tmp_context)
 
     # Call method (will return None, None since no wheels exist)
-    result = bootstrapper._find_cached_wheel(
+    result = _find_cached_wheel(
         bt.ctx,
         bt.cache_wheel_server_url,
         req=Requirement("test-package"),
@@ -274,7 +296,7 @@ def test_get_install_dependencies_returns_list(
     wheel_file = pathlib.Path("/fake/package-1.0.0-py3-none-any.whl")
     unpack_dir = tmp_context.work_dir
 
-    result = bootstrapper._get_install_dependencies(
+    result = _get_install_dependencies(
         ctx=tmp_context,
         req=Requirement("test-package"),
         resolved_version=Version("1.0.0"),
@@ -299,7 +321,7 @@ def test_phase_build_produces_source_build_result(tmp_context: WorkContext) -> N
     mock_sdist_root.parent.mkdir(parents=True, exist_ok=True)
     mock_wheel = tmp_context.work_dir / "package-1.0.0-py3-none-any.whl"
 
-    wi = bootstrapper.WorkItem(
+    wi = WorkItem(
         req=Requirement("test-package"),
         req_type=RequirementType.TOP_LEVEL,
         source_url="https://pypi.org/simple/test-package",
@@ -312,7 +334,7 @@ def test_phase_build_produces_source_build_result(tmp_context: WorkContext) -> N
         build_backend_deps=set(),
         build_sdist_deps=set(),
     )
-    item = bootstrapper.BuildItem(wi)
+    item = BuildItem(wi)
 
     # Set up why stack so _track_why works
     bt.why = []
@@ -325,10 +347,10 @@ def test_phase_build_produces_source_build_result(tmp_context: WorkContext) -> N
             result_items = item.run(bt)
 
     assert len(result_items) == 1
-    assert isinstance(result_items[0], bootstrapper.ProcessInstallDepsItem)
+    assert isinstance(result_items[0], ProcessInstallDepsItem)
 
     result = result_items[0].work_item.build_result
-    assert isinstance(result, bootstrapper.SourceBuildResult)
+    assert isinstance(result, SourceBuildResult)
     assert result.wheel_filename == mock_wheel
     assert result.sdist_filename is None
     assert result.unpack_dir == mock_sdist_root.parent
@@ -355,9 +377,9 @@ def test_multiple_versions_continues_on_error(tmp_context: WorkContext) -> None:
         build_phase_count = {"count": 0}
 
         def prepare_source_run(
-            self: bootstrapper.PrepareSourceItem,
+            self: PrepareSourceItem,
             bt_arg: bootstrapper.Bootstrapper,
-        ) -> list[bootstrapper.PhaseItem]:
+        ) -> list[PhaseItem]:
             build_phase_count["count"] += 1
             if str(self.work_item.resolved_version) == "1.5":
                 raise ValueError("Simulated failure for version 1.5")
@@ -365,7 +387,7 @@ def test_multiple_versions_continues_on_error(tmp_context: WorkContext) -> None:
 
         req = Requirement("testpkg>=1.0")
 
-        with patch.object(bootstrapper.PrepareSourceItem, "run", prepare_source_run):
+        with patch.object(PrepareSourceItem, "run", prepare_source_run):
             with patch.object(bt, "has_been_seen", return_value=False):
                 bt._bootstrap_one(
                     req=req,
@@ -413,7 +435,7 @@ def test_download_wheel_from_cache_bypasses_hooks(
     mock_find_all.side_effect = RuntimeError("no match")
 
     with patch("fromager.overrides.find_and_invoke") as mock_override:
-        result = bootstrapper._download_wheel_from_cache(
+        result = _download_wheel_from_cache(
             bt.ctx,
             bt.cache_wheel_server_url,
             req=Requirement("test-pkg"),
@@ -455,7 +477,7 @@ def test_cache_lookup_resolver_exception_logs_info(
             side_effect=ResolverException("no matching version"),
         ),
     ):
-        result = bootstrapper._download_wheel_from_cache(
+        result = _download_wheel_from_cache(
             bt.ctx,
             bt.cache_wheel_server_url,
             req=Requirement("test-package"),
@@ -487,7 +509,7 @@ def test_cache_lookup_request_exception_logs_warning(
         "fromager.resolver.find_all_matching_from_provider",
         side_effect=exc_class(exc_msg),
     ):
-        result = bootstrapper._download_wheel_from_cache(
+        result = _download_wheel_from_cache(
             bt.ctx,
             bt.cache_wheel_server_url,
             req=Requirement("test-package"),
@@ -509,7 +531,7 @@ def test_cache_lookup_unexpected_exception_logs_warning(
         "fromager.resolver.find_all_matching_from_provider",
         side_effect=ValueError("unexpected parsing error"),
     ):
-        result = bootstrapper._download_wheel_from_cache(
+        result = _download_wheel_from_cache(
             bt.ctx,
             bt.cache_wheel_server_url,
             req=Requirement("test-package"),
@@ -552,15 +574,15 @@ def test_cache_lookup_download_wheel_error_logs_warning(
             ],
         ),
         patch(
-            "fromager.bootstrapper.wheels.extract_info_from_wheel_file",
+            "fromager.wheels.extract_info_from_wheel_file",
             return_value=("test_package", "1.0.0", None, None),
         ),
         patch(
-            "fromager.bootstrapper.wheels.download_wheel",
+            "fromager.wheels.download_wheel",
             side_effect=exc_class(exc_msg),
         ),
     ):
-        result = bootstrapper._download_wheel_from_cache(
+        result = _download_wheel_from_cache(
             bt.ctx,
             bt.cache_wheel_server_url,
             req=Requirement("test-package"),
@@ -576,7 +598,7 @@ def test_cache_lookup_no_cache_url_returns_none(tmp_context: WorkContext) -> Non
     bt = bootstrapper.Bootstrapper(tmp_context)
     bt.cache_wheel_server_url = ""
 
-    result = bootstrapper._download_wheel_from_cache(
+    result = _download_wheel_from_cache(
         bt.ctx,
         bt.cache_wheel_server_url,
         req=Requirement("test-package"),
@@ -591,9 +613,9 @@ def _make_resolve_item(
     req_type: RequirementType = RequirementType.TOP_LEVEL,
     why_snapshot: list[tuple[RequirementType, Requirement, Version]] | None = None,
     parent: tuple[Requirement, Version] | None = None,
-) -> bootstrapper.ResolveItem:
-    return bootstrapper.ResolveItem(
-        bootstrapper.WorkItem(
+) -> ResolveItem:
+    return ResolveItem(
+        WorkItem(
             req=Requirement(req),
             req_type=req_type,
             why_snapshot=why_snapshot or [],
@@ -603,7 +625,7 @@ def _make_resolve_item(
 
 
 def _record_and_load(
-    bt: bootstrapper.Bootstrapper, stack: list[bootstrapper.PhaseItem]
+    bt: bootstrapper.Bootstrapper, stack: list[PhaseItem]
 ) -> list[typing.Any]:
     bt._record_stack_state(stack)
     return typing.cast(list[typing.Any], json.loads(bt._stack_filename.read_text()))
@@ -617,7 +639,7 @@ def test_record_stack_state_minimal_item(tmp_context: WorkContext) -> None:
     result = contents[0]
     assert result["req"] == "testpkg"
     assert result["req_type"] == str(RequirementType.TOP_LEVEL)
-    assert result["phase"] == str(bootstrapper.BootstrapPhase.RESOLVE)
+    assert result["phase"] == str(BootstrapPhase.RESOLVE)
     assert result["resolved_version"] is None
     assert result["source_url"] is None
     assert result["build_sdist_only"] is False
@@ -635,8 +657,8 @@ def test_record_stack_state_full_item(tmp_context: WorkContext) -> None:
     parent_version = Version("2.0")
     why_snapshot = [(RequirementType.INSTALL, parent_req, parent_version)]
 
-    item = bootstrapper.BuildItem(
-        bootstrapper.WorkItem(
+    item = BuildItem(
+        WorkItem(
             req=Requirement("child-pkg>=1.0"),
             req_type=RequirementType.INSTALL,
             why_snapshot=why_snapshot,
@@ -672,8 +694,8 @@ def test_record_stack_state_full_item(tmp_context: WorkContext) -> None:
 def test_record_stack_state_dep_sets_are_sorted(tmp_context: WorkContext) -> None:
     """Mixed-order dep sets come out alphabetically sorted."""
     bt = bootstrapper.Bootstrapper(tmp_context)
-    item = bootstrapper.BuildItem(
-        bootstrapper.WorkItem(
+    item = BuildItem(
+        WorkItem(
             req=Requirement("mypkg"),
             req_type=RequirementType.TOP_LEVEL,
             why_snapshot=[],
@@ -692,7 +714,7 @@ def test_record_stack_state_dep_sets_are_sorted(tmp_context: WorkContext) -> Non
 def test_record_stack_state_writes_file(tmp_context: WorkContext) -> None:
     """File is created; list length matches stack size."""
     bt = bootstrapper.Bootstrapper(tmp_context)
-    stack: list[bootstrapper.PhaseItem] = [
+    stack: list[PhaseItem] = [
         _make_resolve_item("pkga"),
         _make_resolve_item("pkgb"),
     ]
@@ -708,7 +730,7 @@ def test_record_stack_state_writes_file(tmp_context: WorkContext) -> None:
 def test_record_stack_state_ordering(tmp_context: WorkContext) -> None:
     """Index 0 = stack[-1] (next to pop); last index = stack[0]."""
     bt = bootstrapper.Bootstrapper(tmp_context)
-    stack: list[bootstrapper.PhaseItem] = [
+    stack: list[PhaseItem] = [
         _make_resolve_item("pkga"),
         _make_resolve_item("pkgb"),
         _make_resolve_item("pkgc"),
@@ -743,7 +765,7 @@ def test_bootstrap_calls_record_stack_state(tmp_context: WorkContext) -> None:
 
     original = bt._record_stack_state
 
-    def counting_record(stack: list[bootstrapper.PhaseItem]) -> None:
+    def counting_record(stack: list[PhaseItem]) -> None:
         call_count["n"] += 1
         original(stack)
 
@@ -756,7 +778,7 @@ def test_bootstrap_calls_record_stack_state(tmp_context: WorkContext) -> None:
             "resolve",
             return_value=[("https://pypi.test/testpkg-1.0.tar.gz", Version("1.0"))],
         ),
-        patch.object(bootstrapper.StartItem, "run", return_value=[]),
+        patch.object(StartItem, "run", return_value=[]),
     ):
         bt._bootstrap_one(req=req, req_type=RequirementType.TOP_LEVEL)
 
@@ -766,7 +788,7 @@ def test_bootstrap_calls_record_stack_state(tmp_context: WorkContext) -> None:
 def test_bootstrap_with_empty_list(tmp_context: WorkContext) -> None:
     """bootstrap([]) completes without error and runs no phases."""
     bt = bootstrapper.Bootstrapper(tmp_context)
-    with patch.object(bootstrapper.ResolveItem, "run") as mock_run:
+    with patch.object(ResolveItem, "run") as mock_run:
         bt.bootstrap([])
     mock_run.assert_not_called()
 
@@ -775,11 +797,11 @@ def test_bootstrap_with_single_requirement(tmp_context: WorkContext) -> None:
     """bootstrap([req]) resolves and processes the requirement."""
     bt = bootstrapper.Bootstrapper(tmp_context)
     req = Requirement("testpkg==1.0")
-    captured: list[bootstrapper.ResolveItem] = []
+    captured: list[ResolveItem] = []
 
     def capture_run(
-        self: bootstrapper.ResolveItem, bt_arg: bootstrapper.Bootstrapper
-    ) -> list[bootstrapper.PhaseItem]:
+        self: ResolveItem, bt_arg: bootstrapper.Bootstrapper
+    ) -> list[PhaseItem]:
         captured.append(self)
         return []
 
@@ -789,13 +811,13 @@ def test_bootstrap_with_single_requirement(tmp_context: WorkContext) -> None:
             "_resolve_and_add_top_level",
             return_value=("http://example.test/testpkg-1.0.tar.gz", Version("1.0")),
         ),
-        patch.object(bootstrapper.ResolveItem, "run", capture_run),
+        patch.object(ResolveItem, "run", capture_run),
         patch.object(bt, "_record_stack_state"),
     ):
         bt.bootstrap([req])
 
     assert len(captured) == 1
-    assert isinstance(captured[0], bootstrapper.ResolveItem)
+    assert isinstance(captured[0], ResolveItem)
     assert captured[0].work_item.req == req
     assert captured[0].work_item.req_type == RequirementType.TOP_LEVEL
 
@@ -807,7 +829,7 @@ def test_bootstrap_skips_failed_resolution(tmp_context: WorkContext) -> None:
 
     with (
         patch.object(bt, "_resolve_and_add_top_level", return_value=None),
-        patch.object(bootstrapper.ResolveItem, "run") as mock_run,
+        patch.object(ResolveItem, "run") as mock_run,
         patch.object(bt, "_record_stack_state"),
     ):
         bt.bootstrap([req])
@@ -824,8 +846,8 @@ def test_bootstrap_two_requirements_both_processed(tmp_context: WorkContext) -> 
     dispatch_calls: list = []
 
     def capture_run(
-        self: bootstrapper.ResolveItem, bt_arg: bootstrapper.Bootstrapper
-    ) -> list[bootstrapper.PhaseItem]:
+        self: ResolveItem, bt_arg: bootstrapper.Bootstrapper
+    ) -> list[PhaseItem]:
         dispatch_calls.append(self.work_item.req.name)
         return []
 
@@ -835,7 +857,7 @@ def test_bootstrap_two_requirements_both_processed(tmp_context: WorkContext) -> 
             "_resolve_and_add_top_level",
             return_value=("http://example.test/pkg-1.0.tar.gz", Version("1.0")),
         ),
-        patch.object(bootstrapper.ResolveItem, "run", capture_run),
+        patch.object(ResolveItem, "run", capture_run),
         patch.object(bt, "_record_stack_state"),
     ):
         bt.bootstrap([req1, req2])
@@ -858,7 +880,7 @@ def test_bg_prepare_source_log_prefix_includes_version(
         with (
             caplog.at_level(logging.INFO, logger="fromager.bootstrapper"),
             patch(
-                "fromager.bootstrapper._find_cached_wheel",
+                "fromager.bootstrapper._cache._find_cached_wheel",
                 return_value=(None, None),
             ),
             patch(
@@ -871,7 +893,7 @@ def test_bg_prepare_source_log_prefix_includes_version(
             ),
             log.req_ctxvar_context(req, version),
         ):
-            bootstrapper._bg_prepare_source(
+            _bg_prepare_source(
                 ctx=tmp_context,
                 cache_wheel_server_url=None,
                 req=req,
@@ -883,7 +905,7 @@ def test_bg_prepare_source_log_prefix_includes_version(
             messages = [
                 r.getMessage()
                 for r in caplog.records
-                if r.name == "fromager.bootstrapper"
+                if r.name.startswith("fromager.bootstrapper")
             ]
     finally:
         logging.setLogRecordFactory(old_factory)
@@ -915,7 +937,7 @@ def test_bg_prepare_prebuilt_log_prefix_includes_version(
             patch("fromager.server.update_wheel_mirror"),
             log.req_ctxvar_context(req, version),
         ):
-            bootstrapper._bg_prepare_prebuilt(
+            _bg_prepare_prebuilt(
                 ctx=tmp_context,
                 req=req,
                 req_type=RequirementType.INSTALL,
@@ -927,7 +949,7 @@ def test_bg_prepare_prebuilt_log_prefix_includes_version(
             messages = [
                 r.getMessage()
                 for r in caplog.records
-                if r.name == "fromager.bootstrapper"
+                if r.name.startswith("fromager.bootstrapper")
             ]
     finally:
         logging.setLogRecordFactory(old_factory)
@@ -943,7 +965,7 @@ def test_build_item_build_sdist_finds_existing(tmp_context: WorkContext) -> None
     """BuildItem._build_sdist returns cached sdist when finders.find_sdist hits."""
     sdist_root = tmp_context.work_dir / "testpkg-1.0" / "testpkg-1.0"
     sdist_root.mkdir(parents=True, exist_ok=True)
-    wi = bootstrapper.WorkItem(
+    wi = WorkItem(
         req=Requirement("testpkg"),
         req_type=RequirementType.TOP_LEVEL,
         why_snapshot=[],
@@ -951,7 +973,7 @@ def test_build_item_build_sdist_finds_existing(tmp_context: WorkContext) -> None
         sdist_root_dir=sdist_root,
         build_env=Mock(),
     )
-    item = bootstrapper.BuildItem(wi)
+    item = BuildItem(wi)
     cached = tmp_context.sdists_builds / "testpkg-1.0.tar.gz"
     cached.touch()
 
@@ -967,7 +989,7 @@ def test_build_item_build_sdist_calls_build_when_not_cached(
     """BuildItem._build_sdist calls sources.build_sdist when no cached sdist found."""
     sdist_root = tmp_context.work_dir / "testpkg-1.0" / "testpkg-1.0"
     sdist_root.mkdir(parents=True, exist_ok=True)
-    wi = bootstrapper.WorkItem(
+    wi = WorkItem(
         req=Requirement("testpkg"),
         req_type=RequirementType.TOP_LEVEL,
         why_snapshot=[],
@@ -975,7 +997,7 @@ def test_build_item_build_sdist_calls_build_when_not_cached(
         sdist_root_dir=sdist_root,
         build_env=Mock(),
     )
-    item = bootstrapper.BuildItem(wi)
+    item = BuildItem(wi)
     built = tmp_context.sdists_builds / "testpkg-1.0.tar.gz"
 
     with (
@@ -998,7 +1020,7 @@ def test_build_item_build_wheel(tmp_context: WorkContext) -> None:
     """BuildItem._build_wheel builds sdist then wheel and updates mirror."""
     sdist_root = tmp_context.work_dir / "testpkg-1.0" / "testpkg-1.0"
     sdist_root.mkdir(parents=True, exist_ok=True)
-    wi = bootstrapper.WorkItem(
+    wi = WorkItem(
         req=Requirement("testpkg"),
         req_type=RequirementType.TOP_LEVEL,
         why_snapshot=[],
@@ -1006,7 +1028,7 @@ def test_build_item_build_wheel(tmp_context: WorkContext) -> None:
         sdist_root_dir=sdist_root,
         build_env=Mock(),
     )
-    item = bootstrapper.BuildItem(wi)
+    item = BuildItem(wi)
     built_wheel = tmp_context.wheels_build / "testpkg-1.0-py3-none-any.whl"
     built_sdist = tmp_context.sdists_builds / "testpkg-1.0.tar.gz"
 
@@ -1030,7 +1052,7 @@ def test_build_item_do_build_returns_cached_wheel(
     """BuildItem.do_build returns cached wheel immediately without building."""
     cached = tmp_context.wheels_downloads / "testpkg-1.0-py3-none-any.whl"
     cached.touch()
-    wi = bootstrapper.WorkItem(
+    wi = WorkItem(
         req=Requirement("testpkg"),
         req_type=RequirementType.TOP_LEVEL,
         why_snapshot=[],
@@ -1040,7 +1062,7 @@ def test_build_item_do_build_returns_cached_wheel(
         sdist_root_dir=tmp_context.work_dir,
         build_env=Mock(),
     )
-    item = bootstrapper.BuildItem(wi)
+    item = BuildItem(wi)
 
     with patch.object(item, "_build_wheel") as mock_wheel:
         wheel, sdist = item.do_build(tmp_context)
@@ -1055,7 +1077,7 @@ def test_build_item_do_build_sdist_only(tmp_context: WorkContext) -> None:
     built_sdist = tmp_context.sdists_builds / "testpkg-1.0.tar.gz"
     sdist_root = tmp_context.work_dir / "testpkg-1.0" / "testpkg-1.0"
     sdist_root.mkdir(parents=True, exist_ok=True)
-    wi = bootstrapper.WorkItem(
+    wi = WorkItem(
         req=Requirement("testpkg"),
         req_type=RequirementType.TOP_LEVEL,
         why_snapshot=[],
@@ -1065,7 +1087,7 @@ def test_build_item_do_build_sdist_only(tmp_context: WorkContext) -> None:
         sdist_root_dir=sdist_root,
         build_env=Mock(),
     )
-    item = bootstrapper.BuildItem(wi)
+    item = BuildItem(wi)
 
     with (
         patch.object(item, "_build_sdist", return_value=built_sdist) as mock_sdist,
@@ -1085,7 +1107,7 @@ def test_build_item_do_build_builds_wheel(tmp_context: WorkContext) -> None:
     built_sdist = tmp_context.sdists_builds / "testpkg-1.0.tar.gz"
     sdist_root = tmp_context.work_dir / "testpkg-1.0" / "testpkg-1.0"
     sdist_root.mkdir(parents=True, exist_ok=True)
-    wi = bootstrapper.WorkItem(
+    wi = WorkItem(
         req=Requirement("testpkg"),
         req_type=RequirementType.TOP_LEVEL,
         why_snapshot=[],
@@ -1095,7 +1117,7 @@ def test_build_item_do_build_builds_wheel(tmp_context: WorkContext) -> None:
         sdist_root_dir=sdist_root,
         build_env=Mock(),
     )
-    item = bootstrapper.BuildItem(wi)
+    item = BuildItem(wi)
 
     with patch.object(
         item, "_build_wheel", return_value=(built_wheel, built_sdist)
